NEJM Clinical Trials Dataset

A collection of 39 clinical trials published in the New England Journal of Medicine, used to compare Bayesian analysis methods. Each row is a separate trial with binary outcomes.

Usage

nejm

Format

A data frame with 39 rows and 10 variables:

No.

Trial reference number.

P.value

Frequentist p-value.

BF01

Bayes factor in favor of H0.

y1

Number of events in group 1 (control).

y2

Number of events in group 2 (treatment).

n1

Number of subjects in group 1 (control).

n2

Number of subjects in group 2 (treatment).

y1n1

Event rate in group 1 (y1/n1).

y2n2

Event rate in group 2 (y2/n2).

ynavg

Average event rate across both groups.

References

Irons, N.J. and Cinelli, C. (2025). Causally Sound Priors for Binary Experiments. Bayesian Analysis. doi:10.1214/25-BA1506

Examples

data(nejm)
head(nejm)
#>   No. P.value   BF01  y1  y2   n1   n2        y1n1       y2n2       ynavg
#> 1  28    0.62 130.80  26  31 2976 3009 0.008736559 0.01030243 0.009519493
#> 2  28    0.47 110.50  26  33 3032 3077 0.008575198 0.01072473 0.009649965
#> 3  17    0.19  12.30  18  10  493  488 0.036511156 0.02049180 0.028501480
#> 4  52    0.48  85.34 164 151 4996 5007 0.032826261 0.03015778 0.031492020
#> 5  28    0.93  96.00 131 132 3906 3862 0.033538146 0.03417918 0.033858664
#> 6  28    0.93  95.70 141 137 3995 3950 0.035294118 0.03468354 0.034988831